requirements.txt
# LocalCloud - Python Dependencies

# Async web framework (server)
quart>=0.19,<1.0
hypercorn>=0.17,<1.0

# Cryptography
PyNaCl>=1.5,<2.0
argon2-cffi>=23.1,<24.0

# Serialization
cbor2>=5.6,<6.0

# HTTP client (client-side)
httpx>=0.27,<1.0

# CLI
click>=8.1,<9.0

# Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0

# Rust key management crate (built via maturin)
# Install with: cd rust/keycore && maturin develop
